+++ common/src/main/java/fi/hel/verkkokauppa/common/constants/PaymentGatewayEnum.java
@@ -1,0 +1,25 @@
+package fi.hel.verkkokauppa.common.constants;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum PaymentGatewayEnum {
+    VISMA("online"),
+    PAYTRAIL("online-paytrail"),
+    INVOICE("offline");
+
+    @JsonValue
+    private final String type;
+
+    PaymentGatewayEnum(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    @Override
+    public String toString() {
+        return type;
+    }
+}
