+++ orderapi/src/main/java/fi/hel/verkkokauppa/order/repository/jpa/OrderPaymentMethodRepository.java
@@ -1,10 +1,12 @@
-package fi.hel.verkkokauppa.payment.repository;
+package fi.hel.verkkokauppa.order.repository.jpa;
 
-import fi.hel.verkkokauppa.payment.model.OrderPaymentMethod;
+import fi.hel.verkkokauppa.order.model.OrderPaymentMethod;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+@Repository
 public interface OrderPaymentMethodRepository extends ElasticsearchRepository<OrderPaymentMethod, String> {
     List<OrderPaymentMethod> findByOrderId(String orderId);
     long deleteByOrderId(String orderId);
