+++ orderapi/src/main/resources/application-local.properties
@@ -10,13 +10,13 @@ sap.sftp.server.username=sftp
 sap.sftp.server.password=password
 
 kafka.client.authentication.enabled=false
-serviceconfiguration.url=http://host.docker.internal:8187/serviceconfiguration/
-order.experience.url=http://host.docker.internal:8084/v1/order/
+serviceconfiguration.url=http://localhost:8187/serviceconfiguration/
+order.experience.url=http://localhost:8084/v1/order/
 
-history.service.url=http://host.docker.internal:8189
-message.service.url=http://host.docker.internal:8181
-productmapping.service.url=http://host.docker.internal:8187/productmapping/
-merchant.service.url=http://host.docker.internal:8188
+history.service.url=http://localhost:8189
+message.service.url=http://localhost:8181
+productmapping.service.url=http://localhost:8187/productmapping/
+merchant.service.url=http://localhost:8188
 
 spring.activemq.broker-url=failover:(tcp://localhost:61616)
 spring.activemq.user=jmx
